Git Initialisation : First cup of the web interface base on purge JSF without any framework.
In this cup it is possible to manage any of the table except those from user. 
We can CRUD ! Becarful on huge table lazy is not perfectly done ! 

Template is also realize based on OBI frame. 

Intend of this cup is to being able to fill table in order to finalize obi-service colelction restore to normal